Refactor DAO classes and Controllers
